[create]productのダミーデータを、DBに登録

+++ backend/database/migrations/2024_04_07_111202_create_products_table.php
@@ -20,9 +20,9 @@ return new class extends Migration
             $table->string('image');
             $table->string('category');
             $table->text('description');
-            $table->decimal('rating', 5, 2)->default(0);
+            $table->decimal('rating', 3, 1)->default(0);
             $table->unsignedInteger('num_reviews')->default(0);
-            $table->decimal('price', 8, 2)->default(0);
+            $table->integer('price')->default(0);
             $table->enum('contract_type', ['月契約', '単発契約']);
             $table->timestamps();
         });
